requirements.txt
asgiref==3.7.2
certifi==2023.7.22
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.2.0
cloudinary==1.33.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.1.0
Django==4.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.4.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
pdfrw==0.4
phonenumbers==8.13.18
Pillow==9.5.0
psycopg2-binary==2.9.6
pycryptodome==3.18.0
PyJWT==2.8.0
pyotp==2.8.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
rave-python==1.4.0
reportlab==4.2.2
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.16
user-agents==2.2.0
uuid==1.30
whitenoise==6.5.0
